# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Communication_LTD Ticket System API
  description: |
    API for Communication_LTD customer support ticket system
    
    ## Logging Requirements
    
    - All successful mutating operations (POST, PUT, PATCH, DELETE) with valid CSRF tokens are logged at INFO level
    - Missing or invalid CSRF tokens are logged at WARNING level
    - Requests approaching rate limit thresholds are logged at WARNING level
    - All requests include client IP, user ID (if authenticated), timestamp, and operation details
    
    ## Rate Limiting
    
    The API implements rate limiting to protect against abuse:
    - Authentication endpoints: 10 requests per minute per IP
    - Standard endpoints: 60 requests per minute per authenticated user
    - Admin endpoints: 30 requests per minute per admin user
    
    When requests approach 80% of their limit threshold, WARNING level logs are generated
  version: 1.0.0
  contact:
    name: Communication_LTD Support
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      description: >
        CSRF token for mutation requests protection.
        All POST, PUT, PATCH, DELETE operations require this header.
        Missing or invalid tokens will be logged at WARNING level.
        Valid tokens on successful mutating requests will be logged at INFO level.
      required: true
      schema:
        type: string
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
      required:
        - code
        - message
      example:
        code: 400
        message: Bad Request
        details:
          field: "username"
          reason: "Username is already taken"
...
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        address:
          type: string
        city:
          type: string
        postal_code:
          type: string
        country:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - first_name
        - last_name
        - phone_number
        - created_at
        - updated_at
      example:
        id: 1
        user_id: 1
        first_name: "John"
        last_name: "Doe"
        phone_number: "+1-555-123-4567"
        address: "123 Main St"
        city: "Springfield"
        postal_code: "12345"
        country: "USA"
        created_at: "2025-08-01T10:00:00Z"
        updated_at: "2025-08-01T10:00:00Z"
    CustomerCreate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        address:
          type: string
        city:
          type: string
        postal_code:
          type: string
        country:
          type: string
      required:
        - first_name
        - last_name
        - phone_number
      example:
        first_name: "John"
        last_name: "Doe"
        phone_number: "+1-555-123-4567"
        address: "123 Main St"
        city: "Springfield"
        postal_code: "12345"
        country: "USA"
...
paths:
...
  /customers:
    get:
      summary: List all customers
      operationId: listCustomers
      tags:
        - Customers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Number of items per page
        - name: search
          in: query
          schema:
            type: string
          description: Search term for filtering customers
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer
                  limit:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new customer
      operationId: createCustomer
      tags:
        - Customers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customers/{customer_id}:
    parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get customer details
      operationId: getCustomer
      tags:
        - Customers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update customer details
      operationId: updateCustomer
      tags:
        - Customers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a customer
      operationId: deleteCustomer
      tags:
        - Customers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Customer deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...